Introduce submodules to orderbook
